requirements.txt
streamlit
python-dotenv
langchain==0.0.142
pydantic==1.10.7
